Move TypeManager logic to visitor class
