+++ src/Bicep.Core/SemanticModel/ResourceSymbol.cs
@@ -17,37 +17,6 @@ namespace Bicep.Core.SemanticModel
 
         public ResourceDeclarationSyntax DeclaringResource => (ResourceDeclarationSyntax) this.DeclaringSyntax;
 
-        public override TypeSymbol Type
-        {
-            get
-            {
-                // if type string is malformed, the type value will be null which will resolve to a null type
-                // below this will be corrected into an error type
-                var stringSyntax = this.DeclaringResource.TryGetType();
-                TypeSymbol? resourceType;
-
-                if (stringSyntax != null && stringSyntax.IsInterpolated())
-                {
-                    // TODO: in the future, we can relax this check to allow interpolation with compile-time constants.
-                    // right now, codegen will still generate a format string however, which will cause problems for the type.
-                    resourceType = new ErrorTypeSymbol(DiagnosticBuilder.ForPosition(this.DeclaringResource.Type).ResourceTypeInterpolationUnsupported());
-                }
-                else
-                {
-                    var stringContent = stringSyntax?.GetLiteralValue();
-                    resourceType = this.Context.TypeManager.GetTypeByName(stringContent);
-
-                    // TODO: This check is likely too simplistic
-                    if (resourceType?.TypeKind != TypeKind.Resource)
-                    {
-                        resourceType = new ErrorTypeSymbol(DiagnosticBuilder.ForPosition(this.DeclaringResource.Type).InvalidResourceType());
-                    }
-                }
-
-                return resourceType;
-            }
-        }
-
         public SyntaxBase Body { get; }
 
         public override void Accept(SymbolVisitor visitor) => visitor.VisitResourceSymbol(this);
